fix: fix WebSocket poll() from blocking
0a248ff (feat: :zap: reformat Uri; add tests & examples; remove macros;, 2022-10-19)

Previously the timeout values used in SslClient.cpp did not influence its behavior, until the inclusion of the query() function in both send() and receive(). This results in WebSocket blocking on poll(), preventing any future writes to the server. This has been fixed by setting the SslClient to non-blocking for the receive() and resetting it to blocking. Also removed an irrelevant comment on the discord_gateway ws example.