build: :rotating_light: ignore OpenSSL compiler warnings
OpenSSL currently generates a "useless-cast" warning on GCC and an "unused-value" warning which we must ignore to pass the "Werror" flag if enabled. This is completely safe to do as OpenSSL knows what they're doing. Appropriate steps were taken to ensure that ignoring "useless-cast" does not appear on Clang builds which could generate "unknown warning" errors.