build: :hammer: fix libfmt installation
When libfmt is not found on the host system, the current install commands fail and write to the wrong directories on Linux.

This has to do with the way libraries are installed in both OSes.

Windows Libraries store their dependencies within its folder, like the following example:
my-lib
 - bin/
 - include/
 - lib/
Meanwhile, on Linux, libraries are stored in a /usr/local folder in which all libraries' internals are split:
bin/
include/ <- Headers for different installed libraries go here.
lib/

This fix makes it so that Windows systems install libfmt internally, and Linux uses libfmt's bundled install command, which installs correctly.